move glance images to a json directory inside of canned responses
