+++ mimic/canned_responses/json/__init__.py
@@ -1,0 +1,3 @@
+"""
+Canned responses
+"""
