+++ mimic/canned_responses/json/glance/__init__.py
@@ -1,0 +1,3 @@
+"""
+Canned responses
+"""
