+++ mimic/canned_responses/json/glance/glance_images_json.py
